Added Nathan's LCD & Hex keypad functions. Removed all player 1/2 stuff, and SPI stuff.
