+++ README.md
@@ -1,4 +1,10 @@
 # HCS12Battleship
 Battleship project for ECE470
 Chris Taylor, Nathan Burns, Aleksander Danielewski
-Dragon12+ based project. Using Freescale code warrior IDE, Adafruit neopixel matrix, nrf2401 transcievers
+Dragon12+ based project. Using Freescale code warrior IDE, Adafruit neopixel matrix
+
+
+Battleship is a prototype for the project, anything that dosent depend on both players will be prototyped there
+player1 is the master and first player to attack, player2 is the slave and first player to defend
+
+Both player1 and player2's sources are from battleship's sources 3-31 9:49am
